Spells: Add few target checks to MIT spell cast check function Spells: Rename legacy spell .cpp files Spells: Move SpellEffects.h and SpellFailure.h to Spell/Definitions/ Spells: Declare spell and aura effects implemented in Cataclysm Spells: Revert "AuraStates should not be checked for triggered spells" Experimental Filesystem: Correct config directory
