+++ src/world/Management/AchievementMgr.cpp
@@ -29,7 +29,7 @@
 #include "Objects/Faction.h"
 #include "Spell/Definitions/SpellMechanics.h"
 #include "Spell/Customization/SpellCustomizations.hpp"
-#include "Spell/SpellEffects.h"
+#include "Spell/Definitions/SpellEffects.h"
 #include "Guild.h"
 
 #if VERSION_STRING > TBC
