+++ src/world/Spell/CMakeLists.txt
@@ -8,24 +8,22 @@ set(SRC_SPELL_FILES
    ${PATH_PREFIX}/Spell.Legacy.cpp
    ${PATH_PREFIX}/Spell.Legacy.h
    ${PATH_PREFIX}/Spell_ClassScripts.cpp
-   ${PATH_PREFIX}/SpellAuras.cpp
+   ${PATH_PREFIX}/SpellAuras.Legacy.cpp
    ${PATH_PREFIX}/SpellAuras.h
    ${PATH_PREFIX}/SpellCastTargets.h
    ${PATH_PREFIX}/SpellCastTargets.cpp
    ${PATH_PREFIX}/SpellDefines.hpp
-   ${PATH_PREFIX}/SpellEffects.cpp
-   ${PATH_PREFIX}/SpellEffects.h
-   ${PATH_PREFIX}/SpellFailure.h
+   ${PATH_PREFIX}/SpellEffects.Legacy.cpp
    ${PATH_PREFIX}/SpellHelpers.h
    ${PATH_PREFIX}/SpellInfo.cpp
    ${PATH_PREFIX}/SpellInfo.hpp
-   ${PATH_PREFIX}/SpellMgr.cpp
+   ${PATH_PREFIX}/SpellMgr.Legacy.cpp
    ${PATH_PREFIX}/SpellMgr.h
-   ${PATH_PREFIX}/SpellProc.cpp
+   ${PATH_PREFIX}/SpellProc.Legacy.cpp
    ${PATH_PREFIX}/SpellProc.h
    ${PATH_PREFIX}/SpellProc_ClassScripts.cpp
    ${PATH_PREFIX}/SpellProc_Items.cpp
-   ${PATH_PREFIX}/SpellTarget.cpp
+   ${PATH_PREFIX}/SpellTarget.Legacy.cpp
    ${PATH_PREFIX}/SpellTarget.h
    ${PATH_PREFIX}/SpellTargetConstraint.cpp
    ${PATH_PREFIX}/SpellTargetConstraint.h
