+++ src/world/Spell/Definitions/CMakeLists.txt
@@ -21,7 +21,9 @@ set(SRC_SPELL_DEFINITIONS_FILES
     ${PATH_PREFIX}/SpellCoefficientsFlags.h
     ${PATH_PREFIX}/SpellDamageType.h
     ${PATH_PREFIX}/SpellDidHitResult.h
+    ${PATH_PREFIX}/SpellEffects.h
     ${PATH_PREFIX}/SpellEffectTarget.h
+    ${PATH_PREFIX}/SpellFailure.h
     ${PATH_PREFIX}/SpellGoFlags.h
     ${PATH_PREFIX}/SpellInFrontStatus.h
     ${PATH_PREFIX}/SpellIsFlags.h
