+++ src/world/Spell/SpellAuras.Legacy.cpp
