+++ src/world/Spell/SpellEffects.Legacy.cpp
