+++ src/world/Spell/SpellInfo.cpp
@@ -7,7 +7,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Management/Skill.h"
 #include "../../scripts/Battlegrounds/AlteracValley.h"
 #include "Definitions/SpellEffectTarget.h"
-#include "Spell/SpellEffects.h"
+#include "Spell/Definitions/SpellEffects.h"
 #include "Spell/SpellAuras.h"
 
 SpellInfo::SpellInfo()
