+++ src/world/Spell/SpellMgr.Legacy.cpp
