+++ src/world/Spell/SpellProc.Legacy.cpp
