+++ src/world/Spell/SpellTarget.Legacy.cpp
