+++ src/world/StdAfx.h
@@ -463,8 +463,8 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Spell/SpellAuras.h"
 #include "Spell/SpellCastTargets.h"
 #include "Spell/SpellDefines.hpp"
-#include "Spell/SpellEffects.h"
-#include "Spell/SpellFailure.h"
+#include "Spell/Definitions/SpellEffects.h"
+#include "Spell/Definitions/SpellFailure.h"
 #include "Spell/SpellHelpers.h"
 #include "Spell/SpellInfo.hpp"
 #include "Spell/SpellMgr.h"
